src/requirements.txt
# chromedriver_binary==119.0.6022.0.0
chromedriver_binary_auto
#chromedriver_binary==117.0.5938.92
# chromedriver_binary==118.0.5993.70
Flask==2.3.3
flask_cors==4.0.0
flask_sqlalchemy==3.1.1
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.1
pmdarima==2.0.3
pymysql==1.1.0
python-dotenv==1.0.0
Requests==2.31.0
rich==13.5.3
scikit_learn==1.3.1
selenium==4.12.0
ShopifyAPI==12.3.0
SQLAlchemy==2.0.21
statsmodels==0.14.0
webdriver_manager==4.0.0
mysql-connector-python==8.1.0
fuzzywuzzy==0.18.0
Levenshtein==0.22.0
apscheduler==3.10.4
flask-apscheduler==1.13.0
inflect==7.0.0
unidecode==1.3.7